nonexternalprocess/requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.5.2
aniso8601==8.0.0
anyjson==0.3.3
astroid==2.3.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5
billiard==3.5.0.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
DateTime==4.3
docutils==0.15.2
hiredis==1.0.1
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.2.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
multidict==4.7.5
mysql==0.0.2
mysqlclient==1.4.5
protobuf==3.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-engineio==3.11.2
python-ldap==3.2.0
python-socketio==4.4.0
python3-ldap==0.9.8.4
pytz==2019.3
redis==2.10.6
requests==2.22.0
rsa==4.0
s3transfer==0.3.0
six==1.12.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
vine==1.3.0
websocket-client==0.57.0
websockets==8.0.2
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.4.2
zipp==1.0.0
zope.interface==4.7.1
